requirements.txt
gunicorn
Flask
Flask-Compress
Flask-Cors
python-dotenv
joblib
numpy
scikit-learn
lightgbm
catboost